refactor: move state data transitions from extract to their own transition modules
